Filenames are now `pathshorten`d, similar to Fish shell.
Old filename display in statusline:
  /home/freddy/dotfiles/vim/plugged/vim-moonfly/plugin/moonfly.vim

New filename display:
  /h/f/d/v/p/v/p/moonfly.vim

This should prevent most instances of scrolled off filenames when a
window is split many ways. The last component (moonfly.vim above) is the
most important BUT with the shortened path being a handy tool to
disambiguate between files of the same name but in different
directories.

Also refactored Cmdline handling.