Leaving filename in statusline (even when global)
Statusline may have lint errors and Gitsigns; the elimination of the
filename to the statusline sometimes meant that my eye needed to wonder
back to the active window to figure which file is current.

I thought it would be nice to not repeat the filename in multiple
places, but turns out that was not a good idea.