Emphasis color is now dynamic and auto-generated
Git Branch & Total Lines previously were coded to a single color (often
blue). Now it is simple the mode background color, but now as the text
color. So when going from normal to insert to visual mode the emphasized
text will also change in color mirroring the mode indicator area.