Prevent selecting GitGutter when its loaded but not enabled
When the following configuration:
  mistflyWithGitsignsStatus
  mistflyWithGitGutterStatus
  mistflyWithSignifyStatus
was replaced with:
  mistflyWithGitStatus
control was lost over selection of the Git status plugin.

That selecting ability was useful when a vim configuration has multiple
Git status plugins loaded with only one enabled.

This commit allows GitGutter to be loaded but not enabled, and Signify
to be selected (if loaded and enabled).