Remove call to Signify buffer_is_active()
Since Signify is the last Git status plugin in the selecting if-else
statements, checking if Signify is active for the buffer is unnecessary.

Further, since Signify returns [-1,-1,-1] when the buffer is inactive,
the result is the same, i.e., no counts will be displayed.

Removing the call to buffer_is_active() is more efficient.